Restructure render files
