+++ src/ft_mlx/ft_mlx_render.c
@@ -6,12 +6,11 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/17 22:03:01 by rdesprez          #+#    #+#             */
-/*   Updated: 2025/08/07 18:16:46 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2025/09/08 10:58:29 by rdesprez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d_render.h"
-#include "mlx.h"
 
 void	cubmlx_clear(t_mlx *mlx, unsigned int color)
 {
@@ -60,5 +59,51 @@ void	cubmlx_putvertline(t_data *data, t_pos2 pos, int len,
 		*(unsigned int *)px = color;
 		px += step;
 		y++;
+	}
+}
+
+void	cubmlx_putsquare(t_data *data, t_pos2 pos, int size,
+	unsigned int color)
+{
+	int	x;
+	int	y;
+
+	y = 0;
+	while (y < size)
+	{
+		x = 0;
+		while (x < size)
+		{
+			cubmlx_putpixel(data, pos.x + x, pos.y + y, color);
+			x++;
+		}
+		y++;
+	}
+}
+
+void	cubmlx_putline(t_data *data, t_pos2 p1, t_pos2 p2, unsigned int color)
+{
+	t_vec2	delta;
+	t_vec2	pos;
+	int		length;
+	int		i;
+
+	if (absf(p2.x - p1.x) > absf(p2.y - p1.y))
+		length = absf(p2.x - p1.x);
+	else
+		length = absf(p2.y - p1.y);
+	delta.x = (float)(p2.x - p1.x) / length;
+	delta.y = (float)(p2.y - p1.y) / length;
+	pos.x = p1.x + 0.5f;
+	pos.y = p1.y + 0.5f;
+	i = 0;
+	while (i <= length)
+	{
+		if (pos.x >= 0 && pos.x < data->mlx.win_size.x && pos.y >= 0
+			&& pos.y < data->mlx.win_size.y)
+			cubmlx_putpixel(data, (int)pos.x, (int)pos.y, color);
+		pos.x += delta.x;
+		pos.y += delta.y;
+		i++;
 	}
 }
